Handle the case where the reviewer isn't a user
Projects can also be assigned as reviewers of a revision, which
we currently don't handle. This commit adds proper handling for
projects, and also adds a fallback name in the case that fails as
well. This should cover the common cases and also more gracefully
fail when we hit something weird.